move plot logic from CrossValidationItem to CrossValidationReporter
