+++ skore/tests/unit/item/test_cross_validation_item.py
@@ -15,6 +15,7 @@ class TestCrossValidationItem:
                 estimator=None,
                 X=None,
                 y=None,
+                plot_bytes=None,
             )
 
     def test_factory(self, monkeypatch, mock_nowstr):
@@ -35,12 +36,14 @@ class TestCrossValidationItem:
             estimator=MyEstimator(),
             X=[[1.0]],
             y=[1],
+            plot_bytes=b"",
         )
 
         assert item.cv_results_serialized == {"test_score": [1, 2, 3]}
         assert item.estimator_info == {"name": "MyEstimator", "params": "{}"}
         assert item.X_info == {"nb_cols": 1, "nb_rows": 1, "hash": ""}
         assert item.y_info == {"hash": ""}
+        assert item.plot_bytes == b""
         assert isinstance(item.plot_bytes, bytes)
         assert item.created_at == mock_nowstr
         assert item.updated_at == mock_nowstr
